requirements.txt
requests==2.31.0
ddddocr==1.5.5
python-dotenv==1.0.0
